backend_py/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
motor==3.5.1
pydantic==2.8.2
email-validator==2.2.0
jinja2==3.1.4
aiosmtplib==3.0.1
google-generativeai>=0.7.0
